Pushed Missing files to the repository.
This UWP application uses a video device, so it is necessary to enable
Capabilities ('webcam' and 'microphone') of the UWP application.
I forgot to upload 'Package.appxmanifest', so build application is
missing Capabilities , and you couldn't execute
Because the application did not have enough authority, it failed
initialization of 'OcrService'.